Override the render of block based widgets to control rendering
Store the widget arguments, so we send the appropriate wraps to the
render of the menu.

Using a block in widgets will wrap the contents of block
in a `<section>` tag, regardless of the content within.
We mimic the functionality of the inner echo while excluding
the calls to output the wrap.

[#625625414772799]