Provide unique widget ids for category widget multiple widget rendering
When the options is set to `widget` for category widgets, we get duplicate
widget wraps when multiple top level categories are rendered.

Leave the CSS classes and JS selectors intact and change the HTML id to be
unique.

Prevent duplicate rendering of styles and accordion JS within the same widget.

Task [#1207223812785575]